+++ src/commands/link.ts
@@ -15,8 +15,8 @@
  */
 
 import { Command, flags } from '@oclif/command';
-import { AnnounceService, BootstrapService, CommandUtils, LoggerFactory, LogType } from '../';
-import { LinkService } from '../service';
+import { LoggerFactory, LogType } from '../logger';
+import { AnnounceService, BootstrapService, CommandUtils, LinkService } from '../service';
 
 export default class Link extends Command {
     static description = `It announces VRF and Voting Link transactions to the network for each node with 'Peer' or 'Voting' roles. This command finalizes the node registration to an existing network.`;