+++ src/commands/start.ts
@@ -16,7 +16,7 @@
 
 import { Command } from '@oclif/command';
 import { LoggerFactory } from '../logger';
-import { BootstrapService, CommandUtils, Constants } from '../service';
+import { BootstrapAccountResolver, BootstrapService, CommandUtils, Constants } from '../service';
 import Clean from './clean';
 import Compose from './compose';
 import Config from './config';
@@ -48,6 +48,9 @@ export default class Start extends Command {
             CommandUtils.passwordPromptDefaultMessage,
             true,
         );
-        await new BootstrapService(logger).start({ ...flags, workingDir: Constants.defaultWorkingDir });
+
+        const workingDir = Constants.defaultWorkingDir;
+        const accountResolver = new BootstrapAccountResolver(logger);
+        await new BootstrapService(logger).start({ ...flags, accountResolver, workingDir });
     }
 }