+++ test/service/BootstrapServce.test.ts
@@ -16,7 +16,7 @@
 
 import { expect } from 'chai';
 import 'mocha';
-import { Assembly, BootstrapService, Constants, LoggerFactory, LogType, Preset, StartParams } from '../../src';
+import { Assembly, BootstrapService, Constants, DefaultAccountResolver, LoggerFactory, LogType, Preset, StartParams } from '../../src';
 
 const logger = LoggerFactory.getLogger(LogType.Silent);
 describe('BootstrapService', () => {
@@ -29,6 +29,7 @@ describe('BootstrapService', () => {
             reset: true,
             upgrade: false,
             timeout: 60000 * 5,
+            accountResolver: new DefaultAccountResolver(),
             target: 'target/tests/BootstrapService.standard',
             detached: true,
             user: 'current',
@@ -51,6 +52,7 @@ describe('BootstrapService', () => {
             reset: true,
             upgrade: false,
             timeout: 60000 * 5,
+            accountResolver: new DefaultAccountResolver(),
             target: 'target/tests/BootstrapService.testnet',
             detached: true,
             user: 'current',