+++ test/service/BootstrapUtils.test.ts
@@ -20,8 +20,9 @@ import 'mocha';
 import { it } from 'mocha';
 import { totalmem } from 'os';
 import { Account, NetworkType } from 'symbol-sdk';
+import { ConfigurationUtils } from '../../src';
 import { ConfigAccount } from '../../src/model';
-import { BootstrapUtils, ConfigLoader, CryptoUtils } from '../../src/service';
+import { BootstrapUtils, CryptoUtils } from '../../src/service';
 
 describe('BootstrapUtils', () => {
     it('BootstrapUtils generate random', async () => {
@@ -31,7 +32,7 @@ describe('BootstrapUtils', () => {
 
         for (let i = 0; i < 10; i++) {
             console.log();
-            const account = ConfigLoader.toConfig(Account.generateNewAccount(networkType));
+            const account = ConfigurationUtils.toConfigAccount(Account.generateNewAccount(networkType));
             balances.push({ ...account, balance: 1000000 });
         }
         console.log(BootstrapUtils.toYaml({ nemesisBalances: balances }));