+++ test/service/ComposeService.test.ts
@@ -246,6 +246,19 @@ ${BootstrapUtils.toYaml(dockerCompose)}
         await assertDockerCompose(params, 'expected-docker-compose-bootstrap-custom.yml');
     });
 
+    it('Compose mainnet custom services and logging', async () => {
+        const params = {
+            ...ConfigService.defaultParams,
+            ...LinkService.defaultParams,
+            target: 'target/tests/ComposeService-mainnet-custom-services',
+            customPreset: './test/unit-test-profiles/custom_compose_service.yml',
+            reset: true,
+            preset: Preset.testnet,
+            assembly: Assembly.dual,
+        };
+        await assertDockerCompose(params, 'expected-mainnet-custom-services.yml');
+    });
+
     it('Compose bootstrap demo with debug on', async () => {
         const params = {
             ...ConfigService.defaultParams,
@@ -254,7 +267,9 @@ ${BootstrapUtils.toYaml(dockerCompose)}
                 dockerComposeDebugMode: true,
                 faucets: [
                     {
-                        environment: { FAUCET_PRIVATE_KEY: 'MockMe', NATIVE_CURRENCY_ID: 'Mockme2' },
+                        compose: {
+                            environment: { FAUCET_PRIVATE_KEY: 'MockMe', NATIVE_CURRENCY_ID: 'Mockme2' },
+                        },
                     },
                 ],
             },
@@ -287,7 +302,9 @@ ${BootstrapUtils.toYaml(dockerCompose)}
             customPresetObject: {
                 faucets: [
                     {
-                        environment: { FAUCET_PRIVATE_KEY: 'MockMe', NATIVE_CURRENCY_ID: 'Mockme2' },
+                        compose: {
+                            environment: { FAUCET_PRIVATE_KEY: 'MockMe', NATIVE_CURRENCY_ID: 'Mockme2' },
+                        },
                     },
                 ],
             },